Merge main into V4 branch

+++ sdk/src/Services/S3/Custom/_bcl/IO/S3DirectoryInfo.cs
@@ -192,7 +192,7 @@ namespace Amazon.S3.IO
                     ((Amazon.Runtime.Internal.IAmazonWebServiceRequest)request).AddBeforeRequestHandler(S3Helper.FileIORequestEventHandler);
 
                     var response = s3Client.ListObjects(request);
-                    return response.S3Objects.Count > 0;
+                    return response.S3Objects?.Count > 0;
                 }
             }
             catch (AmazonS3Exception e)
@@ -250,12 +250,16 @@ namespace Amazon.S3.IO
                         var listRequest = new ListBucketsRequest();
                         ((Amazon.Runtime.Internal.IAmazonWebServiceRequest)listRequest).AddBeforeRequestHandler(S3Helper.FileIORequestEventHandler);
 
-                        foreach (S3Bucket s3Bucket in s3Client.ListBuckets(listRequest).Buckets)
+                        var buckets = s3Client.ListBuckets(listRequest).Buckets;
+                        if (buckets != null)
                         {
-                            DateTime currentBucketLastWriteTime = new S3DirectoryInfo(s3Client, s3Bucket.BucketName, String.Empty).LastWriteTime;
-                            if (currentBucketLastWriteTime > ret)
+                            foreach (S3Bucket s3Bucket in buckets)
                             {
-                                ret = currentBucketLastWriteTime;
+                                DateTime currentBucketLastWriteTime = new S3DirectoryInfo(s3Client, s3Bucket.BucketName, String.Empty).LastWriteTime;
+                                if (currentBucketLastWriteTime > ret)
+                                {
+                                    ret = currentBucketLastWriteTime;
+                                }
                             }
                         }
                     }
@@ -488,19 +492,21 @@ namespace Amazon.S3.IO
                 {
                     listResponse = s3Client.ListObjects(listRequest);
 
-                    // Sort to make sure the Marker for paging is set to the last lexiographical key.
-                    foreach (S3Object s3o in listResponse.S3Objects.OrderBy(x => x.Key))
+                    if (listResponse.S3Objects != null)
                     {
-                        deleteRequest.AddKey(s3o.Key);
-                        if (deleteRequest.Objects.Count == Util.S3Constants.MULTIPLE_OBJECT_DELETE_LIMIT)
+                        // Sort to make sure the Marker for paging is set to the last lexiographical key.
+                        foreach (S3Object s3o in listResponse.S3Objects.OrderBy(x => x.Key))
                         {
-                            s3Client.DeleteObjects(deleteRequest);
-                            deleteRequest.Objects.Clear();
-                        }
+                            deleteRequest.AddKey(s3o.Key);
+                            if (deleteRequest.Objects.Count == Util.S3Constants.MULTIPLE_OBJECT_DELETE_LIMIT)
+                            {
+                                s3Client.DeleteObjects(deleteRequest);
+                                deleteRequest.Objects.Clear();
+                            }
 
-                        listRequest.Marker = s3o.Key;
+                            listRequest.Marker = s3o.Key;
+                        }
                     }
-                    
                 } while (listResponse.IsTruncated.GetValueOrDefault());
 
                 if (deleteRequest.Objects.Count > 0)
@@ -567,8 +573,17 @@ namespace Amazon.S3.IO
             {
                 var request = new ListBucketsRequest();
                 ((Amazon.Runtime.Internal.IAmazonWebServiceRequest)request).AddBeforeRequestHandler(S3Helper.FileIORequestEventHandler);
-                folders = s3Client.ListBuckets(request).Buckets
-                    .ConvertAll(s3Bucket => new S3DirectoryInfo(s3Client,s3Bucket.BucketName,""));
+
+                var buckets = s3Client.ListBuckets(request).Buckets;
+                if (buckets != null)
+                {
+                    folders = buckets
+                        .ConvertAll(s3Bucket => new S3DirectoryInfo(s3Client, s3Bucket.BucketName, ""));
+                }
+                else
+                {
+                    folders = new List<S3DirectoryInfo>();
+                }
             }
             else
             {
@@ -1208,7 +1223,7 @@ namespace Amazon.S3.IO
             do
             {
                 var buckets = this.S3Client.ListBuckets().Buckets;
-                currentState = buckets.FirstOrDefault(x => string.Equals(x.BucketName, this.BucketName)) != null;
+                currentState = buckets?.FirstOrDefault(x => string.Equals(x.BucketName, this.BucketName)) != null;
 
                 if (currentState == exists)
                 {
